+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+2.3.0 / 2018-05-03
+  * Engine.removeFact() - removes fact from the engine (@SaschaDeWaal)
+  * Engine.removeRule() - removes rule from the engine (@SaschaDeWaal)
+  * Engine.removeOperator() - removes operator from the engine (@SaschaDeWaal)
+
+2.2.0 / 2018-04-19
+  * Performance: Constant facts now perform 18-26X better
+  * Performance: Removes await/async transpilation and json.stringify calls, significantly improving overall performance
+
+2.1.0 / 2018-02-19
+  * Publish dist updates for 2.0.3
+
+2.0.3 / 2018-01-29
+  * Add factResult and result to the JSON generated for Condition (@bjacobso)
+
+2.0.2 / 2017-07-24
+  * Bugfix IE8 support
+
 2.0.1 / 2017-07-05
   * Bugfix rule result serialization